Check for duplicate constant name conflicts
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1922 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ rexxapi/windows/RexxAPIManager.c
@@ -634,7 +634,7 @@ int   FillAPIComBlock(
 
 int RxGetModAddress(const char *dll_name,
                     char *function_name,      // gets uppercased potentially, so can't be const
-                    size_t *error_codes,
+                    int  *error_codes,
                     REXXPFN * function_address)
 {
    int      rc = 0 ;                   /* Function result.           */
